Fix functional tests
- Add config to use sauce labs instead of python virtual display
- Remove python 3.3 as it is not supported by saucelabs
