+++ .travis.yml
@@ -3,13 +3,12 @@ services:
   - mongodb
 python:
   - "2.7"
-  - "3.4"
+#  - "3.4"
 # command to install dependencies
-#before_install:
-#    - 'export XVFBARGS=":99 -ac -screen 0 1024x768x16"XVFBA'
-#    - "export DISPLAY=:99.0"
-#    - "sh -e /etc/init.d/xvfb start"
-#    - sleep 3 # give xvfb some time to start
+addons:
+    sauce_connect:
+        username: "echaloasuerte"
+        access_key: "966c755a-b4b0-42f9-80ec-5e5a0f0375ab"
 install:
   - "pip install -r requirements.txt"
   - export DJANGO_SETTINGS_MODULE="echaloasuerte.settings.test"
@@ -17,4 +16,4 @@ install:
 # command to run tests
 script:
   - nosetests test_init.py server/test -v
-  - nosetests func_tests -v
+  - nosetests func_test -v
